fixing db now

+++ Restaurant.BL/Models/Food.cs
@@ -5,7 +5,7 @@ namespace Restaurant.BL.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public float Weight { get; set; }
-        public float CookingTime { get; set; }
+        public int CookingTime { get; set; }
         public string Ingredients { get; set; }
         public bool FoodNeedInstrument { get; set; }   
     }
