+++ Restaurant.DAL/Entities/FoodEntity.cs
@@ -4,7 +4,7 @@ namespace Restaurant.DAL.Entities
     {
         public string Name { get; set; }
         public float Weight { get; set; }
-        public float CookingTime { get; set; }
+        public int CookingTime { get; set; }
         public string Ingredients { get; set; }
         public bool FoodNeedInstrument { get; set; }
     }
